+++ activity/brightpay/react/changelog.md
@@ -1980,3 +1980,14 @@
 
 ---
 
+# Brightpay Development - 2025-01-11
+
+**Commit #19**
+
+- Repository: brightpay/brightpay
+- Type: web-application
+- Language: react
+- Month: 1/2025
+
+---
+